% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{apply_mask}
\alias{apply_mask}
\title{Apply a mask to a Nifti file or array}
\usage{
apply_mask(image, mask, return_type = c("flattened", "volume"))
}
\arguments{
\item{image}{A 3- or 4-dimensional array to be masked}

\item{mask}{Mask to be applied to the image. Must be created using \link{create_mask}.}

\item{return_type}{either "flattened" for a matrix result or "volume" if the original volumetric structure should be preserved. Default is "flattened".}
}
\description{
Apply a mask to a Nifti file or array
}
