% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbss.R
\name{bbss}
\alias{bbss}
\title{Run Bayesian blind source separation on prewhitened fMRI time courses}
\usage{
bbss(
  Y,
  initial_guess,
  formula = NULL,
  data = NULL,
  maxit = 100,
  Hmax = 10,
  Hmax_Rqv = 5,
  concentration = 1,
  concentration_Rqv = 1,
  site_variable = NULL,
  coding_scheme = "effects",
  base_measure_lambda = 1,
  base_measure_mu = 0,
  base_measure_alpha = 2,
  base_measure_beta = 1,
  print_freq = 5,
  store_subject_maps = FALSE
)
}
\arguments{
\item{Y}{The prewhitened fMRI time courses. Each time course should be prewhitened to \eqn{Q} timepoints, where \eqn{Q} is the number of spatial source signals (components) to be estimated. See details below.}

\item{initial_guess}{Starting values for some of the model parameters. At a minimum, this must include an initial value for the population average brain network map in a field called \code{init_S} that is of dimension \eqn{Q \times V}, where \eqn{Q} is the number of maps to estimate and \eqn{V} is the number of spatial locations. See details below.}

\item{formula}{A \link{formula} object specifying the covariate effects and random terms (in the case of longitudinal data). Follows the same structure as the formula from a \link{lm} or \link{lmer} function. See details below.}

\item{data}{A \link{tibble} or \link{data.frame} containing the covariate effects. Anything in the \code{formula} argument must also have a corresponding column in data.}

\item{maxit}{Number of iterations to run when optimizing the parameters of the variational distribution. Note that this currently does not involve early stopping (DEV VERSION).}

\item{Hmax}{Maximum number of clusters in the DPM for the spatial source signals}

\item{Hmax_Rqv}{Maximum number of clusters in the DPM for the covariance of the random effects}

\item{concentration}{Concentration parameter for the DPM for the spatial source signals}

\item{concentration_Rqv}{Concentration parameter for the DPM for the covariance of the random effects}

\item{site_variable}{Name of the column of \code{data} containing the site variable indicators. If this is included, a site-adjusted BBSS model will be fit that incorporates random intercepts for site.}

\item{coding_scheme}{Specify the coding scheme to apply to any categorical variables in the data. Options are "effects" coding (DEFAULT) or "reference" cell coding (IN DEV)}

\item{base_measure_lambda}{Hyperparameter for lambda in the base measure for the covariance of the random effects}

\item{base_measure_mu}{Hyperparameter for the mean in the base measure for the covariance of the random effects. Default is 0.}

\item{base_measure_alpha}{Hyperparameter for the shape in the base measure for the covariance of the random effects}

\item{base_measure_beta}{Hyperparameter for scale in the base measure for the covariance of the random effects}

\item{print_freq}{How often summary information about the algorithm should be printed to the console. Default is every 5 iterations.}

\item{store_subject_maps}{Boolean variable storing whether the individual subject-level map estimates should be stored. Default is FALSE, since these can use up a substantial amount of memory.}
}
\value{
An object of class BBSSModel containing parameter estimates, fitted spatial maps, covariate effects, and algorithm diagnostic information.
}
\description{
Run Bayesian blind source separation on prewhitened fMRI time courses
}
\details{
The function fits a variational approximation to the SparseBayes ICA (Lukemire, 2023)  and REMBRAiNDT models (under review). Models for \code{bbss} are specified symbolically, similar to \link{lm} and \link{lmer}.
}
\section{Preparing the \code{Y} Argument}{

\code{bbss} expects the \code{Y} argument to contain the prewhitened time courses for each participant and scan (in the event of multiple scans per subject). The function will accept either a list or array argument for \code{Y} in the following forms:

\itemize{
  \item A 3D Array that is \eqn{Q \times V \times N^*}, where \eqn{N^*} is the total number of scans across all subjects. This is equivalent to concatenating the data along the 3rd dimension. Note that for datasets without repeated measures this is just \eqn{Q \times V \times N}.
  \item A 4D Array that is \eqn{Q \times V \times N \times J^*}, where \eqn{J^*} is the maximum number of scans any subject has. Note that this can be memory inefficient if some subjects have many scans and others only have a few.
  \item A list of length \eqn{N}, where list element \eqn{i} is another list containing \eqn{J_i} matrices of dimension \eqn{Q \times V}. 
  \item A list of length \eqn{N}, where list element \eqn{i} is an array of size \eqn{Q \times V \times J_i}.
}

The prewhitening can be carried out using the \link{prewhiten_participant} function provided by this library. The most convenient format will depend on how your data is stored and whether or not you have multiple scans per subject. Finally, the internal functions that parse the input will attempt to fix some orientation errors, for example if a \eqn{V \times Q \times N} array is provided, the library will handle permuting the indices as needed.
}

\section{Preparing Initial Guess}{

An initial guess for the model parameters must be provided in the \code{initial_guess} argument. The \code{initial_guess} argument should be structured as a list with the following elements:

\itemize{
  \item \code{init_S}: (Required) A matrix (\eqn{Q \times V} or \eqn{V \times Q}) containing the initial values for the spatial maps (components). This can be obtained using FASTICA via the \link{obtain_spatial_map_guess_fastica} function provided in this library. \bold{This is the only required element of the \code{initial_guess} list}. 
  \item \code{init_Ai}: (Optional) An array (\eqn{Q \times Q \times N^*}) containing the initial values for each subject's mixing matrix. Note that if this is not provided, \link{bbss} will calculate a guess for you based on a regression of the prewhitened time courses onto \code{init_S}.
}

The remaining parameters will be initialized within the \link{bbss} function.
}

\section{Preparing the \code{data} Argument}{

The \code{data} argument should be either a \link{tibble} or \link{data.frame} containing the values of any clinical or demographic covariates. This should be prepared in the same way as the \code{data} argument from \link{lm} or \link{lmer}. It can also be useful to include a column containing the filepath to each nifti file, as this can be used to help ensure that the order of \code{Y} is the same as the order of the covariates by applying the preprocessing one-row-at-a-time using this column.
}

\section{Specifying the model formula}{

The model \code{formula} should be provided with no term on the left-hand side. Some example forms:
\itemize{
  \item \code{~ sex + age}: Will fit a model with two estimated beta maps: one for sex and one for age. Requires \code{data} to have columns named "sex" and "age".
  \item \code{~ sex*age}: Will fit a model with three estimated beta maps: one for sex, one for age, and one for a sex by age interaction. Requires \code{data} to have columns named "sex" and "age".
  \item \code{~ sex + age + (1 | ID)}: Will fit a model with two estimated beta maps: one for sex and one for age. Additionally, this model will include a random intercept corresponding to \code{ID}. Requires \code{data} to have columns named "sex", "age", and "ID".
}

Note that the column names can be any valid column name, and the above uses of "sex", "age", "ID" are just examples.
}

\references{
Lukemire, J., Pagnoni, G., & Guo, Y. (2023). \emph{Sparse Bayesian modeling of hierarchical independent component analysis: Reliable estimation of individual differences in brain networks}.
Biometrics, 79(4), 3599-3611.
}
\seealso{
\link{prewhiten_participant}, \link{obtain_spatial_map_guess_fastica}
}
