% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_to_reference.R
\name{compare_2d_map_correlations}
\alias{compare_2d_map_correlations}
\title{Compare two sets of spatial maps}
\usage{
compare_2d_map_correlations(map1, map2, Q = NULL, cormethod = "pearson")
}
\arguments{
\item{map1}{A spatial map of dimension \eqn{Q \times V}.}

\item{map2}{A spatial map of dimension \eqn{Q \times V} or \eqn{V \times Q}.}

\item{Q}{Optional argument specifying the number of components. Only needed if \code{map1} or \code{map2} are not already \eqn{Q \times V}.}
}
\value{
A list object with the following fields:

\itemize{
  \item \code{correlation_unsorted}: Pearson correlation between the two maps using their raw ordering.
  \item \code{sorted_order}: Vector giving the re-indexing of \code{map1} that makes it most closely match \code{map2}.
  \item \code{correlation_sorted}: Pearson correlation between the two maps after re-indexing.
}
}
\description{
Compare two sets of spatial maps
}
