% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{create_mask}
\alias{create_mask}
\title{Create a mask}
\usage{
create_mask(image, threshold = 0)
}
\arguments{
\item{image}{A nifti image or 3d array to be used to generate the brain mask}

\item{threshold}{Everything greater than this value will be included in the mask}
}
\description{
Create a mask
}
