% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fns.R
\name{generate_model_matrix.BBSSModel}
\alias{generate_model_matrix.BBSSModel}
\title{Generate the model matrix for a BBSSModel based on a new set of covariate values}
\usage{
generate_model_matrix.BBSSModel(obj, newdata)
}
\arguments{
\item{obj}{An object of class \code{"BBSSModel"} output from \link{bbss}.}

\item{newdata}{A \code{data.frame} (or tibble) with the covariate values at which
to construct the model matrix.}
}
\value{
A matrix containing the model matrix based on \code{newdata}.
}
\description{
Generate the model matrix for a BBSSModel based on a new set of covariate values
}
