% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{obtain_spatial_map_guess_fastica}
\alias{obtain_spatial_map_guess_fastica}
\title{Obtain an initial set of spatial components using FastICA}
\usage{
obtain_spatial_map_guess_fastica(data)
}
\arguments{
\item{data}{Numeric matrix of dimension \eqn{Q \times V}, typically the output of 
\link{PCA_dimension_reduction}, where \eqn{Q} is the number of components and 
\eqn{V} is the number of spatial locations (voxels).}
}
\value{
A numeric matrix of dimension \eqn{Q \times V}. Each row corresponds to an 
  estimated spatial component (loadings across voxels) obtained from the FastICA 
  algorithm.
}
\description{
Obtain an initial set of spatial components using FastICA
}
\details{
This function applies FastICA to the reduced data to provide an initial set of spatial 
components. These can be used as starting values for the population-average spatial maps 
in a subsequent \code{bbss} model fit.
}
\seealso{
Related function: \link{PCA_dimension_reduction} for performing PCA prior to ICA.
}
