% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_assessments.R
\name{plot_relative_changes}
\alias{plot_relative_changes}
\title{Plot relative changes at each CAVI iteration for key model parameters}
\usage{
plot_relative_changes(obj, plot_type = "Variance")
}
\arguments{
\item{obj}{an object of \link{class} "BBSSModel"}

\item{plot_type}{Either "Variance" or "DPM", depending on which parameters you wish to view}
}
\value{
A ggplot object
}
\description{
Plot relative changes at each CAVI iteration for key model parameters
}
\details{
This function generates a ggplot object showing relative change over each iteration for either the variance parameters or the DPM parameters, depending on user selection.
}
