% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{prewhiten_participant}
\alias{prewhiten_participant}
\title{Prewhiten fMRI time courses}
\usage{
prewhiten_participant(data, Q)
}
\arguments{
\item{data}{Numeric matrix of fMRI time courses, with \eqn{T} timepoints (rows) and \eqn{V} spatial locations (columns).}

\item{Q}{Integer. Target dimensionality after prewhitening (must match the number of components to be estimated in blind source separation).}
}
\value{
A list with the following elements:
\itemize{
  \item \code{Y}: a \eqn{V \times Q} matrix of prewhitened time courses.
  \item \code{eigenvalues}: numeric vector of eigenvalues of the temporal covariance matrix.
}
}
\description{
Prewhiten fMRI time courses
}
\examples{

# Simulate example time courses (T=50, V=100)
set.seed(100)
simulated_time_course <- matrix(rnorm(100*50), nrow = 50, ncol = 100)

# Prewhiten to Q = 5 components
prewhitened_time_courses <- prewhiten_participant(simulated_time_course, Q = 5)
cov(prewhitened_time_courses$Y)

}
