% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_to_reference.R
\name{reorder_to_reference}
\alias{reorder_to_reference}
\title{Reorder the results of a blind source separation to match a set of reference images}
\usage{
reorder_to_reference(obj, reference_img)
}
\arguments{
\item{obj}{an object of \link{class} "BBSSModel"}

\item{reference_img}{A matrix of dimension \eqn{Q \times V} or \eqn{V \times Q} containing a set of spatial maps to use as a reference for re-ordering the contents of \code{obj}.}
}
\value{
An object of \link{class} "BBSSModel" containing the results stored in \code{obj} re-indexed so that the components are in the order closest matching the components in \code{reference_img}.
}
\description{
Reorder the results of a blind source separation to match a set of reference images
}
\details{
The order of the components in the BSS decomposition is not identifiable, but it is often the case that we want to compare multiple sets of components. Doing so requires re-indexing the components to match as closely as possible. This function calculates the correlation between the components stored in \code{obj} and the maps provided in \code{reference_img}. The any component-specific fields in \code{obj} are then re-ordered to obtain the closest possible matching.
}
